# contracts/generated/search-api.openapi.yaml
components:
...
  schemas:
    AiParseRequest:
      additionalProperties: false
      properties:
        input:
          maxLength: 200
          minLength: 1
          pattern: \S
          type: string
      required:
        - input
      type: object
    AiParseResponse:
      oneOf:
        - additionalProperties: false
          properties:
            amount:
              oneOf:
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - g
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - kg
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - ml
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - piece
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - bar
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - slice
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - portion
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - package
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
            barcode:
              type: "null"
            clarificationQuestion:
              maxLength: 240
              minLength: 1
              type:
                - string
                - "null"
            parser:
              enum:
                - openai
              type: string
            product:
              additionalProperties: false
              properties:
                brand:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
                name:
                  maxLength: 120
                  minLength: 1
                  pattern: \S
                  type: string
                variant:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
              required:
                - name
                - brand
                - variant
              type: object
            rawInput:
              maxLength: 200
              minLength: 1
              pattern: \S
              type: string
            resolutionMode:
              enum:
                - generic_category
              type: string
            status:
              enum:
                - parsed
                - needs_clarification
                - unsupported
              type: string
          required:
            - status
            - rawInput
            - product
            - amount
            - clarificationQuestion
            - parser
            - resolutionMode
            - barcode
          type: object
        - additionalProperties: false
          properties:
            amount:
              oneOf:
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - g
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - kg
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - ml
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - piece
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - bar
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - slice
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - portion
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - package
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
            barcode:
              type: "null"
            clarificationQuestion:
              maxLength: 240
              minLength: 1
              type:
                - string
                - "null"
            parser:
              enum:
                - openai
              type: string
            product:
              additionalProperties: false
              properties:
                brand:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
                name:
                  maxLength: 120
                  minLength: 1
                  pattern: \S
                  type: string
                variant:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
              required:
                - name
                - brand
                - variant
              type: object
            rawInput:
              maxLength: 200
              minLength: 1
              pattern: \S
              type: string
            resolutionMode:
              enum:
                - exact_product
              type: string
            status:
              enum:
                - parsed
                - needs_clarification
                - unsupported
              type: string
          required:
            - status
            - rawInput
            - product
            - amount
            - clarificationQuestion
            - parser
            - resolutionMode
            - barcode
          type: object
        - additionalProperties: false
          properties:
            amount:
              oneOf:
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - g
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - kg
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - ml
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 100000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - piece
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - bar
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - slice
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - portion
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
                - additionalProperties: false
                  properties:
                    unit:
                      enum:
                        - package
                      type: string
                    value:
                      exclusiveMinimum: 0
                      maximum: 1000
                      type: number
                  required:
                    - value
                    - unit
                  type: object
            barcode:
              pattern: ^\d{7,14}$
              type: string
            clarificationQuestion:
              maxLength: 240
              minLength: 1
              type:
                - string
                - "null"
            parser:
              enum:
                - openai
              type: string
            product:
              additionalProperties: false
              properties:
                brand:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
                name:
                  maxLength: 120
                  minLength: 1
                  pattern: \S
                  type: string
                variant:
                  maxLength: 80
                  minLength: 1
                  type:
                    - string
                    - "null"
              required:
                - name
                - brand
                - variant
              type: object
            rawInput:
              maxLength: 200
              minLength: 1
              pattern: \S
              type: string
            resolutionMode:
              enum:
                - barcode
              type: string
            status:
              enum:
                - parsed
                - needs_clarification
                - unsupported
              type: string
          required:
            - status
            - rawInput
            - product
            - amount
            - clarificationQuestion
            - parser
            - resolutionMode
            - barcode
          type: object
    ApiAttemptDiagnostic:
      additionalProperties: false
      properties:
        backend:
          $ref: "#/components/schemas/ApiBackend"
        cacheAgeMs:
          minimum: 0
          type: number
        durationMs:
          minimum: 0
          type: number
        errorMessage:
          type: string
        errorName:
          type: string
        label:
          minLength: 1
          type: string
        outcome:
          $ref: "#/components/schemas/ApiAttemptOutcome"
        responsePreview:
          maxLength: 500
          type: string
        retryAfterMs:
          minimum: 0
          type: number
        startedAt:
          format: date-time
          type: string
        status:
          maximum: 599
          minimum: 100
          type: integer
        url:
          minLength: 1
          type: string
      required:
        - backend
        - label
        - url
        - startedAt
        - durationMs
        - outcome
      type: object
    ApiAttemptOutcome:
      enum:
        - cache-hit
        - success
        - http-error
        - rate-limit
        - network-error
        - timeout
        - parse-error
        - aborted
      type: string
    ApiBackend:
      enum:
        - gateway
        - search-index
        - search-a-licious
        - open-food-facts-legacy
        - open-food-facts-v3
        - open-food-facts-v2
        - query-cache
        - product-cache
      type: string
    ApiError:
      additionalProperties: false
      properties:
        attempts:
          items:
            $ref: "#/components/schemas/ApiAttemptDiagnostic"
          type: array
        code:
          pattern: ^[A-Z][A-Z0-9_]{1,79}$
          type: string
        detail:
          type: string
        error:
          minLength: 1
          type: string
        query_requested:
          type: string
        query_used:
          type: string
        retryAllowedImmediately:
          type: boolean
        retryAt:
          format: date-time
          type: string
        traceId:
          minLength: 8
          type: string
      required:
        - error
        - code
        - traceId
      type: object
...
    ProductApiMode:
      default: hybrid
      enum:
        - hybrid
        - v3
        - v2
      example: hybrid
      type: string
    ProductGatewayResponse:
      additionalProperties: false
      properties:
        api_meta:
          $ref: "#/components/schemas/ApiResponseMeta"
        code:
          pattern: ^\d{7,14}$
          type: string
        gateway_attempts:
          items:
            $ref: "#/components/schemas/ApiAttemptDiagnostic"
          type: array
        product:
          $ref: "#/components/schemas/OffProduct"
        status:
          minLength: 1
          type: string
      required:
        - status
        - code
        - product
        - api_meta
        - gateway_attempts
      type: object
    SearchApiMode:
      default: auto
      description: auto nutzt den eigenen Index primär und OFF Legacy nur als Reserve.
      enum:
        - auto
        - search-index
        - search-a-licious
        - legacy
      example: auto
      type: string
    SearchGatewayResponse:
      additionalProperties: false
      properties:
        api_meta:
          $ref: "#/components/schemas/ApiResponseMeta"
        count:
          minimum: 0
          type: number
        gateway_attempts:
          items:
            $ref: "#/components/schemas/ApiAttemptDiagnostic"
          type: array
        hits:
          items:
            $ref: "#/components/schemas/SearchHit"
          type: array
        page:
          minimum: 1
          type: integer
        page_count:
          minimum: 0
          type: number
        page_size:
          maximum: 20
          minimum: 1
          type: integer
        query_used:
          maxLength: 120
          minLength: 1
          type: string
        source:
          enum:
            - gateway
            - search-index
            - search-a-licious
            - open-food-facts-legacy
            - none
          type: string
        timed_out:
          type: boolean
        took:
          minimum: 0
          type: number
      required:
        - hits
        - api_meta
        - source
        - gateway_attempts
        - query_used
      type: object
...
info:
  description: Kanonischer Vertrag des versionierten Daten-Gateways. Die statische PWA bleibt für manuelle und lokale Funktionen ohne Gateway nutzbar; globale Suche läuft ausschließlich über diese API.
  title: KH Checker data gateway API
  version: 2.2.4
...
openapi: 3.1.0
paths:
  /api/v1/ai/parse:
    post:
      operationId: parseFoodRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiParseRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiParseResponse"
          description: Strukturiertes Parser-Ergebnis
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige Anfrage
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: JSON-Request überschreitet das Gateway-Größenlimit
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Upstream- oder Parser-Limit. Retry-After ist nur ein Hinweis.
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Unerwarteter interner Gateway-Fehler
        "502":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige oder fehlgeschlagene Upstream-Antwort
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Optionaler Dienst vorübergehend nicht verfügbar
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Die absolute Gateway-Deadline wurde überschritten
      summary: Mehrdeutige Nutzereingabe optional strukturieren
      tags:
        - ai
...
  /api/v1/product/{code}:
    get:
      operationId: productByBarcode
      parameters:
        - in: path
          name: code
          required: true
          schema:
            example: "1234567"
            pattern: ^\d{7,14}$
            type: string
        - in: query
          name: known_carbs
          required: false
          schema:
            default: "0"
            enum:
              - "0"
              - "1"
            example: "1"
            type: string
        - in: query
          name: product_api
          required: false
          schema:
            $ref: "#/components/schemas/ProductApiMode"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductGatewayResponse"
          description: Ausgewähltes Produkt
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige Anfrage
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Kein Produkt für diesen Barcode
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: JSON-Request überschreitet das Gateway-Größenlimit
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Upstream- oder Parser-Limit. Retry-After ist nur ein Hinweis.
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Unerwarteter interner Gateway-Fehler
        "502":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige oder fehlgeschlagene Upstream-Antwort
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Optionaler Dienst vorübergehend nicht verfügbar
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Die absolute Gateway-Deadline wurde überschritten
      summary: Nur das ausgewählte Produkt hydratisieren
      tags:
        - product
  /api/v1/search:
    get:
      operationId: searchProducts
      parameters:
        - in: query
          name: q
          required: true
          schema:
            example: BiFi
            maxLength: 120
            minLength: 1
            pattern: \S
            type: string
        - in: query
          name: page_size
          required: false
          schema:
            default: 15
            example: 15
            maximum: 20
            minimum: 1
            type: integer
        - description: auto nutzt den eigenen Index primär und OFF Legacy nur als Reserve.
          in: query
          name: search_api
          required: false
          schema:
            $ref: "#/components/schemas/SearchApiMode"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchGatewayResponse"
          description: Suche abgeschlossen; auch null Treffer sind HTTP 200
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige Anfrage
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: JSON-Request überschreitet das Gateway-Größenlimit
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Upstream- oder Parser-Limit. Retry-After ist nur ein Hinweis.
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Unerwarteter interner Gateway-Fehler
        "502":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Ungültige oder fehlgeschlagene Upstream-Antwort
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Optionaler Dienst vorübergehend nicht verfügbar
          headers:
            Retry-After:
              description: Delta-Sekunden oder HTTP-Datum
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Die absolute Gateway-Deadline wurde überschritten
      summary: Produkte über das explizit konfigurierte Gateway suchen
      tags:
        - search
security: []
servers:
  - description: Same-origin optionales Gateway
    url: /
  - description: Beispiel für ein externes HTTPS-Gateway
    url: https://kh-api.example
...