# src/generated/models/aiParseRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * KH Checker data gateway API
 * Kanonischer Vertrag des versionierten Daten-Gateways. Die statische PWA bleibt für manuelle und lokale Funktionen ohne Gateway nutzbar; globale Suche läuft ausschließlich über diese API.
 * OpenAPI spec version: 2.2.4
 */

export interface AiParseRequest {
  /**
     * @minLength 1
     * @maxLength 200
     * @pattern \S
     */
  input: string;
}